Fix warning about unused macro definition. Closes #272.
The warning was case by `crypto::onetimeauth` reusing the `auth_module`
macro from `crypto::auth`, but not using the `auth_state` macro.

Moved the `auth_state` macro into a separate file to solve this issue.
