+++ src/crypto/auth/mod.rs
@@ -68,6 +68,8 @@
 pub use self::hmacsha512256::*;
 #[macro_use]
 mod auth_macros;
+#[macro_use]
+mod auth_state_macros;
 pub mod hmacsha512;
 pub mod hmacsha512256;
 pub mod hmacsha256;
